requirements.txt
# Conexión a Base de Datos
sqlalchemy>=2.0.0
pyodbc>=5.0.0

# Procesamiento de Datos
pandas>=2.0.0

# Utilidades
python-dotenv>=1.0.0